# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Parking POC API
  version: 0.1.0
  x-contract-status: frozen  # M0-frozen — changes only via reviewed PR
  description: >
    Office parking-slot booking platform. Contract-first source of truth (frozen at M0 — changes via
    reviewed PR). Conventions, envelope, and error codes per docs/phases-3-6-plan.md §3.1–3.2.
    Maturity of each operation is marked with `x-maturity` (DEMO | MVP | Later).
servers:
  - url: /api/v1
    description: Base path (path-based versioning)

# Default: every operation requires a bearer access token. Public endpoints override with `security: []`.
security:
  - bearerAuth: []
...
paths:
...
  /vehicles:
    get:
      tags: [Gate]
      operationId: searchVehicles
      summary: Typeahead over the vehicle registry
      description: >
        Roles SECURITY, SUPER_ADMIN. Matches on the normalized plate or the owner's name. Not
        tenant-scoped (D15) — one guard serves every company in the building.
      x-maturity: MVP
      x-required-roles: [SECURITY, SUPER_ADMIN]
      parameters:
        - { name: search, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 25 } }
      responses:
        '200':
          description: Matching vehicles (empty when `search` is blank).
          headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/VehicleSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /vehicles/lookup:
    get:
      tags: [Gate]
      operationId: lookupVehicle
      summary: Resolve owner + today's booking from a car number
      description: >
        Roles SECURITY, SUPER_ADMIN. The car number is normalized (uppercased, separators stripped),
        so any spacing works. Never 404s for an unrecognised plate — `known: false` is a normal,
        recordable case (D16). Also reports whether the car is already inside (`openVisit`).
      x-maturity: MVP
      x-required-roles: [SECURITY, SUPER_ADMIN]
      parameters:
        - { name: number, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: Lookup result.
          headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/GateLookup' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /gate/events:
    get:
      tags: [Gate]
      operationId: listGateEvents
      summary: Gate log for a date (paged, newest first)
      description: 'Roles SECURITY, SUPER_ADMIN. Defaults to today (IST). Building-wide (D15).'
      x-maturity: MVP
      x-required-roles: [SECURITY, SUPER_ADMIN]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - { name: date, in: query, required: false, schema: { type: string, format: date } }
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [CHECKED_IN, CHECKED_OUT] }
      responses:
        '200':
          description: Paged gate log.
          headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    required: [meta]
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/GateEvent' }
                      meta: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /gate/unbooked:
    get:
      tags: [Gate]
      operationId: listUnbookedEntries
      summary: Entries recorded without a booking (company-admin follow-up feed)
      description: >
        Roles COMPANY_ADMIN (own company) or SUPER_ADMIN (all; optional `companyId`). These are the
        D16 entries the guard let through without a booking. An unregistered plate has no company, so
        it appears only for the SUPER_ADMIN — nobody can attribute an unknown car to a tenant.
      x-maturity: MVP
      x-required-roles: [COMPANY_ADMIN, SUPER_ADMIN]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - { name: date, in: query, required: false, schema: { type: string, format: date } }
        - name: companyId
          in: query
          required: false
          description: SUPER_ADMIN only — narrow to one company. Ignored for COMPANY_ADMIN.
          schema: { type: string }
      responses:
        '200':
          description: Paged unbooked entries.
          headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    required: [meta]
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/GateEvent' }
                      meta: { $ref: '#/components/schemas/Pagination' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ---------------------------------------------------------------------------
  # Reports  (P3-09 — Later)
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token from `POST /auth/login`, sent as `Authorization: Bearer <token>`."
  headers:
    CorrelationId:
      description: Request/response correlation id (echoed into AuditLog).
      schema: { type: string }
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      description: 1-based page number.
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: pageSize
      in: query
      required: false
      description: Items per page (max 100).
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  schemas:
...
    SuccessEnvelope:
      type: object
      description: Standard success envelope. `data` is the resource/list; `meta` present only on list endpoints.
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        # No `type` here: each endpoint overrides `data` (object or array) via allOf.
        data:
          description: Resource or list payload; concrete shape defined per endpoint.
        meta: { $ref: '#/components/schemas/Pagination' }
    ErrorEnvelope:
      type: object
      description: Standard error envelope.
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error: { $ref: '#/components/schemas/ErrorObject' }
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string }
        details:
          type: array
          items: { $ref: '#/components/schemas/ErrorDetail' }
    ErrorDetail:
      type: object
      required: [field, message]
      properties:
        field: { type: string }
        message: { type: string }
    ErrorCode:
      type: string
      description: Canonical error codes (docs/phases-3-6-plan.md §3.2).
      enum:
        - VALIDATION_ERROR   # 400
        - UNAUTHENTICATED    # 401
        - FORBIDDEN          # 403
        - NOT_FOUND          # 404
        - CONFLICT           # 409
        - CAPACITY_FULL      # 409 — every slot for the date is held (Phase 7 D12)
        - WINDOW_CLOSED      # 422
        - RATE_LIMITED       # 429
        - INTERNAL           # 500
    Pagination:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }

    # --- Enums (from prisma/schema.prisma) ---
...
    BookingType:
      type: string
      enum: [PRIMARY, COMMON_POOL]
    BookingStatus:
      type: string
      enum: [DRAFT, SUBMITTED, CANCELLED, ALLOCATED, WAITLISTED, REJECTED, RELEASED, EXPIRED]
    VehicleType:
      type: string
      enum: [CAR, BIKE, EV_CAR, EV_BIKE, OTHER]
...
    VehicleSummary:
      type: object
      required: [id, vehicleNumber, displayNumber, ownerName, vehicleType]
      properties:
        id: { type: string }
        vehicleNumber: { type: string, description: Normalized plate (uppercase, no separators). }
        displayNumber: { type: string, description: Plate as written in the source sheet. }
        ownerName: { type: string }
        ownerEmail: { type: string, nullable: true }
        contactNumber: { type: string, nullable: true }
        vehicleType: { $ref: '#/components/schemas/VehicleType' }
        makeModel: { type: string, nullable: true }
        colour: { type: string, nullable: true }
        companyId: { type: string, nullable: true }
        companyName: { type: string, nullable: true }
    GateLookup:
      type: object
      required: [vehicleNumber, known, bookingDate, hasBooking]
      properties:
        vehicleNumber: { type: string }
        known: { type: boolean, description: False for a plate not in the registry — still checkable-in (D16). }
        vehicle:
          allOf: [ { $ref: '#/components/schemas/VehicleSummary' } ]
          nullable: true
        bookingDate: { type: string, format: date, description: Today's IST business date. }
        hasBooking: { type: boolean }
        booking:
          type: object
          nullable: true
          required: [id, status, bookingType]
          properties:
            id: { type: string }
            status: { $ref: '#/components/schemas/BookingStatus' }
            bookingType: { $ref: '#/components/schemas/BookingType' }
            allocatedSlotNumber: { type: string, nullable: true }
        openVisit:
          type: object
          nullable: true
          description: Set when the car is already inside and has not checked out.
          required: [id, checkInAt]
          properties:
            id: { type: string }
            checkInAt: { type: string, format: date-time }
    GateEventStatus:
      type: string
      enum: [CHECKED_IN, CHECKED_OUT]
    GateEvent:
      type: object
      required: [id, vehicleNumber, displayNumber, bookingDate, hadBooking, status, checkInAt]
      properties:
        id: { type: string }
        vehicleNumber: { type: string }
        displayNumber: { type: string }
        ownerName: { type: string, nullable: true }
        companyId: { type: string, nullable: true }
        companyName: { type: string, nullable: true }
        bookingDate: { type: string, format: date }
        bookingRequestId: { type: string, nullable: true }
        hadBooking:
          type: boolean
          description: False → surfaced on the company-admin dashboard as an entry without a booking (D16).
        status: { $ref: '#/components/schemas/GateEventStatus' }
        checkInAt: { type: string, format: date-time }
        checkOutAt: { type: string, format: date-time, nullable: true }
        notes: { type: string, nullable: true }
...
  responses:
...
    ValidationError:
      description: 400 — request validation failed.
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Request validation failed
              details: [ { field: distanceKm, message: must be between 0 and 200 } ]
    Unauthorized:
      description: 401 — missing/invalid/expired access token or bad credentials.
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { success: false, error: { code: UNAUTHENTICATED, message: Invalid credentials } }
    Forbidden:
      description: 403 — role or tenant-scope violation (or inactive account).
      headers: { X-Correlation-Id: { $ref: '#/components/headers/CorrelationId' } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { success: false, error: { code: FORBIDDEN, message: Account is not active } }
...